Remove the SessionFactory interface
.. and let its only implementing class inherit the name.
